fix: move the collection_info to cloud_service_info
