+++ src/spaceone/inventory/info/__init__.py
@@ -5,7 +5,6 @@ from spaceone.inventory.info.job_task_info import *
 from spaceone.inventory.info.cloud_service_type_info import *
 from spaceone.inventory.info.cloud_service_info import *
 from spaceone.inventory.info.collector_info import *
-from spaceone.inventory.info.collection_info import *
 from spaceone.inventory.info.resource_group_info import *
 from spaceone.inventory.info.change_history_info import *
 from spaceone.inventory.info.note_info import *
